updated baseline.py
